+++ packages/hw-wallet-sdk/src/index.ts
@@ -1,5 +1,6 @@
 export * from './hardware-errors-enums';
 export * from './hardware-error-mappings';
 export * from './hardware-error';
+export * from './dmk-error-mappings';
 export * from './types';
 export * from './user-rejection';